Separate serialization/deserialization tests
